fix duplicate finally
