+++ app/oauth/models.py
@@ -1,20 +1,6 @@
-import random
-import string
 from django.contrib.auth.models import AbstractUser
 from django.db import models
-
-
-def rand_string(length=32):
-    choices = (string.ascii_uppercase + string.ascii_lowercase + string.digits)
-    return ''.join(random.choices(choices, k=length))
-
-
-def rand_color_hex(prefix: str = '#'):
-    rgb = ""
-    for _ in "RGB":
-        i = random.randrange(0, 2**8)
-        rgb += i.to_bytes(1, "big").hex()
-    return f'{prefix or ""}{rgb}'
+from home.util.rand import rand_string, rand_color_hex
 
 
 class CustomUser(AbstractUser):
